Make orders service to divide view and logic
